feat(core)!: exact BigDecimal money paths + vendor error mapping parity
- Move fee estimation to BigDecimal end to end: WalletProvider.estimateTransactionFee
  now returns BigDecimal (breaking for provider implementations); deprecate the lossy
  Double converters in favor of convertWeiToEthDecimal / convertHexToDecimal
- Add strict hex parsing (parseHexToBigIntegerStrict) on balance and gas reads so a
  malformed RPC payload throws instead of silently reading as zero

+++ rain-core-android/src/main/java/com/rain/sdk/internal/provider/WalletProvider.kt
@@ -147,7 +147,7 @@ interface WalletProvider {
      * @param to The target contract address
      * @param data Hex-encoded calldata (or "0x" / empty for plain transfers)
      * @param value Hex-encoded wei value (e.g. "0x0")
-     * @return Estimated fee in the chain's native token
+     * @return Estimated fee in the chain's native token, as an exact [BigDecimal]
      */
     suspend fun estimateTransactionFee(
         chainId: Int,
@@ -155,5 +155,5 @@ interface WalletProvider {
         to: String,
         data: String,
         value: String
-    ): Double
+    ): BigDecimal
 }
